Move the filters from io.kroxylicious.proxy.internal.filter to io.kroxylicious.proxy.filter.simpletransform;
