+++ kroxylicious-filters/kroxylicious-simple-transform/src/main/resources/META-INF/services/io.kroxylicious.proxy.filter.FilterFactory
@@ -1,2 +1,2 @@
-io.kroxylicious.proxy.internal.filter.ProduceRequestTransformationFilterFactory
-io.kroxylicious.proxy.internal.filter.FetchResponseTransformationFilterFactory
+io.kroxylicious.proxy.filter.simpletransform.ProduceRequestTransformationFilterFactory
+io.kroxylicious.proxy.filter.simpletransform.FetchResponseTransformationFilterFactory
