+++ kroxylicious-filters/kroxylicious-simple-transform/src/main/resources/META-INF/services/io.kroxylicious.proxy.filter.simpletransform.ByteBufferTransformationFactory
@@ -4,4 +4,4 @@
 # Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
 #
 
-io.kroxylicious.proxy.internal.filter.UpperCasing
+io.kroxylicious.proxy.filter.simpletransform.UpperCasing
