+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/proxy/FilterIT.java
@@ -56,8 +56,8 @@ import io.kroxylicious.proxy.filter.RejectingCreateTopicFilter;
 import io.kroxylicious.proxy.filter.RejectingCreateTopicFilterFactory;
 import io.kroxylicious.proxy.filter.RequestResponseMarkingFilter;
 import io.kroxylicious.proxy.filter.RequestResponseMarkingFilterFactory;
-import io.kroxylicious.proxy.internal.filter.FetchResponseTransformationFilterFactory;
-import io.kroxylicious.proxy.internal.filter.ProduceRequestTransformationFilterFactory;
+import io.kroxylicious.proxy.filter.simpletransform.FetchResponseTransformationFilterFactory;
+import io.kroxylicious.proxy.filter.simpletransform.ProduceRequestTransformationFilterFactory;
 import io.kroxylicious.test.Request;
 import io.kroxylicious.test.Response;
 import io.kroxylicious.test.ResponsePayload;
