+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/proxy/TestEncoderFactory.java
@@ -8,8 +8,8 @@ package io.kroxylicious.proxy;
 
 import java.nio.ByteBuffer;
 
-import io.kroxylicious.proxy.internal.filter.ByteBufferTransformation;
-import io.kroxylicious.proxy.internal.filter.ByteBufferTransformationFactory;
+import io.kroxylicious.proxy.filter.simpletransform.ByteBufferTransformation;
+import io.kroxylicious.proxy.filter.simpletransform.ByteBufferTransformationFactory;
 import io.kroxylicious.proxy.plugin.Plugin;
 import io.kroxylicious.proxy.plugin.PluginConfigurationException;
