+++ kroxylicious-integration-tests/src/test/resources/META-INF/services/io.kroxylicious.proxy.filter.simpletransform.ByteBufferTransformationFactory
